+++ crates/environ/Cargo.toml
@@ -16,10 +16,7 @@ anyhow = "1.0"
 cranelift-codegen = { path = "../../cranelift/codegen", version = "0.62.0", features = ["enable-serde"] }
 cranelift-entity = { path = "../../cranelift/entity", version = "0.62.0", features = ["enable-serde"] }
 cranelift-wasm = { path = "../../cranelift/wasm", version = "0.62.0", features = ["enable-serde"] }
-wasmparser = "0.51.2"
-lightbeam = { path = "../lightbeam", optional = true, version = "0.15.0" }
 indexmap = "1.0.2"
-rayon = "1.2.1"
 thiserror = "1.0.4"
 directories = "2.0.1"
 sha2 = "0.8.0"
