+++ crates/jit/src/lib.rs
@@ -23,9 +23,14 @@
 
 mod code_memory;
 mod compiler;
+mod cranelift;
+mod func_environ;
 mod imports;
 mod instantiate;
+#[cfg(feature = "lightbeam")]
+mod lightbeam;
 mod link;
+mod module_environ;
 mod resolver;
 mod unwind;
 
@@ -33,9 +38,13 @@ pub mod native;
 pub mod trampoline;
 
 pub use crate::code_memory::CodeMemory;
-pub use crate::compiler::{make_trampoline, Compilation, CompilationStrategy, Compiler};
+pub use crate::compiler::{make_trampoline, Backend, Compilation, CompilationStrategy, Compiler};
+pub use crate::cranelift::Cranelift;
 pub use crate::instantiate::{CompiledModule, SetupError};
+#[cfg(feature = "lightbeam")]
+pub use crate::lightbeam::Lightbeam;
 pub use crate::link::link_module;
+pub use crate::module_environ::ModuleEnvironment;
 pub use crate::resolver::{NullResolver, Resolver};
 
 /// Version number of this crate.
